+++ fetcher/sources/geonode.py
@@ -1,0 +1,49 @@
+# -*- coding: utf-8 -*-
+"""
+-------------------------------------------------
+   File Name：     geonode.py
+   Description :   Geonode代理源
+   Author :        JHao
+   date：          2026/5/31
+-------------------------------------------------
+   Change Activity:
+                   2026/05/31:
+-------------------------------------------------
+"""
+__author__ = 'JHao'
+
+from fetcher.baseFetcher import BaseFetcher
+from handler.logHandler import LogHandler
+from util.webRequest import WebRequest
+
+logger = LogHandler("fetcher")
+
+
+class GeonodeFetcher(BaseFetcher):
+    """Geonode Free Proxy https://geonode.com/free-proxy-list/"""
+
+    name = "geonode"
+    url = "https://geonode.com/free-proxy-list/"
+
+    def fetch(self):
+        url = ("https://proxylist.geonode.com/api/proxy-list?"
+               "limit=500&page=1&sort_by=lastChecked&sort_type=desc")
+        r = WebRequest().get(url, timeout=5, retry_time=1, verify=False)
+        try:
+            proxies = []
+            for item in r.json.get("data", []):
+                ip = item.get("ip", "")
+                port = item.get("port", "")
+                if ip and port:
+                    proxies.append("%s:%s" % (ip, port))
+            if not proxies:
+                proxies = self.parseProxiesFromText(r.text)
+            for proxy in self.yieldUniqueProxies(proxies):
+                yield proxy
+        except Exception as e:
+            logger.error("ProxyFetch - geonode: %s" % e)
+
+
+if __name__ == '__main__':
+    for proxy in GeonodeFetcher().fetch():
+        print(proxy)