+++ proxyPool.py
@@ -39,5 +39,20 @@ def server():
     startServer()
 
 
+@cli.command(name="fetcher")
+def fetcher():
+    """ 查看启用的代理源 """
+    from helper.fetch import _discover_fetchers
+    from handler.configHandler import ConfigHandler
+    conf = ConfigHandler()
+    exclude = conf.fetcherExclude
+    fetcher_classes = _discover_fetchers(exclude)
+    click.echo("Active fetchers (%d):" % len(fetcher_classes))
+    for cls in fetcher_classes:
+        click.echo("  - %s" % cls.name)
+    if exclude:
+        click.echo("\nExcluded: %s" % ", ".join(exclude))
+
+
 if __name__ == '__main__':
     cli()